+++ VacationTracking/AuthenticationServer/Models/ViewModels/AccountViewModel.cs
@@ -31,7 +31,7 @@ namespace AuthenticationServer.Models.ViewModels
 
         public string Avatar { get; set; }
 
-        public int RemainingDaysOff { get; set; }
+        public double RemainingDaysOff { get; set; }
 
         public DateTime? CreatedAt { get; set; }
